Discord/requirements.txt
charset-normalizer==2.0.7
idna==3.3
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
tweepy==4.1.0
urllib3==1.26.7
django
gunicorn
django-heroku
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
discord==1.0.1
discord.py==1.6.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
python-dotenv==0.15.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
webdriver-manager==3.2.2
wrapt==1.12.1
yarl==1.6.3
